% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{show_analysis_modal}
\alias{show_analysis_modal}
\title{Show Detailed Process Analysis Modal}
\usage{
show_analysis_modal(click, res, output)
}
\arguments{
\item{click}{The \code{input$table_cell_clicked} object from DT.}

\item{res}{The \code{processed_info()} reactive list containing the data matrix.}

\item{output}{The Shiny output object to render the plots.}
}
\description{
Triggers a popup containing a trend plot with specification limits and 
a distribution histogram based on a table cell click.
}
