% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_helper.R
\name{ui}
\alias{ui}
\title{Process Capability Analyzer UI}
\description{
Defines the interface for the Cpk Analyzer, including sidebar inputs for file 
uploads and parameter selection, and a main panel for data visualization.
}
\section{UI Components}{


This layout utilizes a Sidebar-Main structure. The sidebar handles data 
ingestion and filtering, while the main panel hosts the interactive 
DataTables and summary statistics.
}

