% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{Process Capability Analyzer Server}
\usage{
server(input, output, session)
}
\description{
Handles the backend logic for data processing and export.
}
\section{Data Ingestion}{

}

\section{Selection Observers}{

}

\section{Processing Logic}{

}

\section{Excel Export}{

}

\section{Table Rendering}{

}

