% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R, R/doc_helper.R
\docType{data}
\name{ui}
\alias{ui}
\title{Process Capability Analyzer UI}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\description{
Defines the interface for the Cpk Analyzer, including sidebar inputs for file 
uploads and parameter selection, and a main panel for data visualization.
}
\section{UI Components}{


This layout utilizes a Sidebar-Main structure. The sidebar handles data 
ingestion and filtering, while the main panel hosts the interactive 
DataTables and summary statistics.
}

\keyword{datasets}
