% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{ui}
\alias{ui}
\title{Process Capability Analyzer UI}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\description{
Defines the layout, sidebar controls, and CSS styling.
}
\keyword{datasets}
